refactor: モックフォーマッタAgMockConstructorをクラスにリファクタリングし、不要なコードを削除
- AgMockConstructor.class.ts:
  AgMockConstructorの型定義および型ガードを追加
- AgMockConstructor.spec.ts: インポートパスを新しいAgMockConstructorクラスの型に更新
- AgMockFormatter.ts: インポートパスを新しいAgMockConstructorクラスの型に更新
- AgMockFormatter.spec.ts: 新しいAgMockConstructorクラスの型を使用するようインポートを更新
